docs: tutorials are moved on the top-level of docs (#4464)
# Added Tutorials section on the top-level of documentation

**Problem Statement**: the Tutorials section in the documentation is
top-priority. Not every project has resources to make tutorials. We have
such a privilege. Community experts created several tutorials on
YouTube.
But the tutorial links are now hidden on the YouTube page and not easily
discovered by first-time visitors.

**PR**: I've created the `Tutorials` page (from the `Additional
Resources/YouTube` page) and moved it to the top level of documentation
in the `Getting Started` section.

## Who can review?

        @dev2049
 
NOTE:
PR checks are randomly failing

https://github.com/hwchase17/langchain/pull/4464/commits/3aefaafcdbb6312f2963163a69874e77d84c63dd

https://github.com/hwchase17/langchain/pull/4464/commits/258819eadfbc45cb0959b187132bd51321ca7370

https://github.com/hwchase17/langchain/pull/4464/commits/514d81b5b3aa12eeed1e8a6eca8d86126697781c