Fix LLM types so that they can be loaded from config dicts (#6235)
LLM configurations can be loaded from a Python dict (or JSON file
deserialized as dict) using the
[load_llm_from_config](https://github.com/hwchase17/langchain/blob/8e1a7a8646dd2e64400c2011fbdcc148127d8ffd/langchain/llms/loading.py#L12)
function.

However, the type string in the `type_to_cls_dict` lookup dict differs
from the type string defined in some LLM classes. This means that the
LLM object can be saved, but not loaded again, because the type strings
differ.